+++ src/app/_components/maintenance.tsx
@@ -1,0 +1,8 @@
+export const Maintenance = () => {
+  return (
+    <div className="mx-auto flex w-full max-w-4xl flex-1 flex-col items-center justify-center px-3 py-10">
+      <p className="text-2xl font-semibold">We&apos;re currently working on a minor update.</p>
+      <span>We will notify you when it&apos;s ready, thanks! ❤️</span>
+    </div>
+  )
+}