+++ src/utils/email/sendEmail.ts
@@ -2,19 +2,20 @@ import { env } from '@/config/env.mjs'
 import { transporter } from '@/config/nodemailer'
 
 export const sendEmail = async ({ to, subject, html }: { to: string; subject: string; html: string }) => {
-  const mailOptions = {
-    from: env.EMAIL_USER,
-    to: to,
-    subject: subject,
-    text: 'PSTrack 🔥',
-    html: html,
-  }
-
   try {
-    const info = await transporter.sendMail(mailOptions)
-    console.log('Email sent:', info.response)
+    const info = await transporter.sendMail({
+      from: env.EMAIL_USER,
+      to: to,
+      subject: subject,
+      html: html,
+    })
+    console.log('To:', to)
     return info
-  } catch (error) {
-    console.error('Error sending email:', error)
+  } catch (error: unknown) {
+    if (error instanceof Error) {
+      console.error('Error sending email to', to, ':', error.stack)
+    } else {
+      console.error('Error sending email to', to, ':', String(error))
+    }
   }
 }