+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
-## [2.2.13]
+## [2.2.15]
 use neetcode 250 roadmap
+
+## [2.1.15]
+update nodemailer config
+
+## [2.1.14]
+add rate limit for emails
 
 ## [2.1.13]
 update roadmap